WIP #367: check dupes dag added to pull_wys dag
